Added key object
When a key is added to a level and activated, it will rotate on the spot.
If player collides with it, it will deactivate and remove itself from the
scene. It will also set the corresponding index of gameObject's levelKeys
to 'true'.

The current plan is to check gameObject's levelKeys when loading a level.
If the key from a level is already picked up, it won't spawn again when
the level loads again.

When the "Overworld" level loads, total number of keys is checked, and
entrances to other levels will open depending on the number of keys
collected (like number of Stars in Mario 64).