Aesthetic update to level one
New skybox (with correct lighting), textures, trees, updated light and
shadow. The coolest thing, though: fog. The fog is not used to simulate
fog, but rather the effect that the sky puts on distant objects (in real
life). That is, the further away an object is from you, the bluer it gets.
The effect is exaggerated here. Making the skybox not be affected by fog
took a lot of research & work.